almost there, need to fix date format
