+++ requirements.txt
@@ -7,3 +7,4 @@ django-contrib-comments==1.6.1
 django-comments-xtd==1.5.0
 django-longerusernameandemail
 requests
+lxml
